Add --global flag to clean task

+++ packages/hardhat-core/src/builtin-tasks/clean.ts
@@ -1,16 +1,23 @@
 import fsExtra from "fs-extra";
 
-import { task } from "../internal/core/config/config-env";
+import { internalTask, task } from "../internal/core/config/config-env";
+import { getCacheDir } from "../internal/util/global-dir";
 
-import { TASK_CLEAN } from "./task-names";
+import { TASK_CLEAN, TASK_CLEAN_GLOBAL } from "./task-names";
 
 export default function () {
-  task(
-    TASK_CLEAN,
-    "Clears the cache and deletes all artifacts",
-    async (_, { config }) => {
+  internalTask(TASK_CLEAN_GLOBAL, async () => {
+    const globalCacheDir = await getCacheDir();
+    await fsExtra.emptyDir(globalCacheDir);
+  });
+
+  task(TASK_CLEAN, "Clears the cache and deletes all artifacts")
+    .addFlag("global", "Clear the global cache")
+    .setAction(async ({ global }: { global: boolean }, { config, run }) => {
+      if (global) {
+        return run(TASK_CLEAN_GLOBAL);
+      }
       await fsExtra.emptyDir(config.paths.cache);
       await fsExtra.remove(config.paths.artifacts);
-    }
-  );
+    });
 }
